+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 2.0.0
+
+* add spring boot 3 support with starter
+* add generic converter
+* move javax.persistence-api to jakarta.persistence-api
+
 # 1.0.1
 
 *  Fixed the destination type to build the mapper.
