--- build.gradle
@@ -1,15 +1,0 @@
-version '1.0.1'
-
-apply plugin: 'java'
-
-sourceCompatibility = 1.8
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    testCompile group: 'junit', name: 'junit', version: '4.12'
-    compile group: 'javax.persistence', name: 'javax.persistence-api', version: '2.2'
-    compile group: 'org.modelmapper', name: 'modelmapper', version: '0.7.5'
-}
