+++ jpa-modelmapper-spring-boot-starter/pom.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+	<parent>
+		<groupId>com.vmf</groupId>
+		<artifactId>jpa-modelmapper</artifactId>
+		<version>2.0.0</version>
+	</parent>
+	<artifactId>jpa-modelmapper-spring-boot-starter</artifactId>
+	<name>jpa-modelmapper-spring-boot-starter</name>
+	<description>jpa-modelmapper Springboot starter</description>
+	<dependencies>
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>com.vmf</groupId>
+			<artifactId>jpa-modelmapper-core</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>org.projectlombok</groupId>
+			<artifactId>lombok</artifactId>
+			<optional>true</optional>
+		</dependency>
+	</dependencies>
+</project>
