+++ jpa-modelmapper-spring-boot-starter/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,1 @@
+com.vmf.modelmapper.jpa.autoconfiguration.JpaModelMapperAutoConfiguration
