--- settings.gradle
@@ -1,2 +1,0 @@
-rootProject.name = 'modelmapper-jpa'
-
